% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringaddins-package.R
\docType{package}
\name{stringaddins-package}
\alias{stringaddins}
\alias{stringaddins-package}
\title{Manipulate Strings}
\description{
RStudio add-ins for manipulating strings.
}
\details{
Open a file in RStudio. Select text and execute one of the add-ins.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hmalmedal/stringaddins}
  \item Report bugs at \url{https://github.com/hmalmedal/stringaddins/issues}
}

}
\author{
\strong{Maintainer}: Håkon Malmedal \email{hmalmedal@gmail.com}

}
\keyword{package}
